requirements.txt
google-adk>=0.1.11
google-generativeai>=0.3.2
uuid>=1.30
argparse>=1.4.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
Pillow>=10.0.0  # For image processing
PyPDF2>=3.0.0   # For PDF processing 